% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spat_abund.R
\name{generate_spat_abund}
\alias{generate_spat_abund}
\title{Generate neutral community}
\usage{
generate_spat_abund(theta, Ivec, Jvec)
}
\arguments{
\item{theta}{a number}

\item{Ivec}{a vector}

\item{Jvec}{a vector}
}
\value{
a vector of species abundances
}
\description{
Generate neutral community
}
\examples{
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000))
}
